requirements.txt
faiss-cpu
streamlit
spacy
numpy
skillNer
unidecode
pdfplumber
wordcloud
matplotlib
re 
collections